Reorganized Layout:
* `src` foder is now `esbmc_ai_lib` folder.
* `esbmc_ai_lib` folder is now a module.
* `commands` folder is now a module.
* Added `esbmc-ai` and `esbmc_ai.py` files that run the project from the
root folder.
