+++ esbmc_ai.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+
+# Author: Yiannis Charalambous 2023
+
+from esbmc_ai_lib.main import main
+
+main()
