+++ esbmc_ai_lib/__init__.py
@@ -1,0 +1,1 @@
+version: str = "0.01"
