+++ esbmc_ai_lib/commands/__init__.py
@@ -1,0 +1,5 @@
+from .chat_command import ChatCommand
+from .exit_command import ExitCommand
+from .fix_code_command import FixCodeCommand
+from .help_command import HelpCommand
+from .verify_code_command import VerifyCodeCommand
