+++ esbmc_ai_lib/commands/chat_command.py
